property-tests/property.go
package main

import (
	"strings"
)

func ConvertToRoman(arabic int) string {
	var result strings.Builder

	for _, numeral := range allRomanNumerals {
		for arabic >= numeral.Value {
			result.WriteString(numeral.Symbol)
			arabic -= numeral.Value
		}
	}
	return result.String()
}

func ConvertToInt(Symbol string) int {

	var result = 0
	for i := 0; i < len(Symbol); i++ {
		if Symbol == "IV" {
			result += 4
			break
		}
	}
	return result
}

//All these values have a maximum degree
//They can all be defined in terms of getting the max degree and then handling the remainder

type RomanNumeral struct {
	Value  int
	Symbol string
}

var allRomanNumerals = []RomanNumeral{
	{1000, "M"},
	{900, "CM"},
	{500, "D"},
	{400, "CD"},
	{100, "C"},
	{50, "L"},
	{40, "XL"},
	{10, "X"},
	{9, "IX"},
	{5, "V"},
	{4, "IV"},
	{3, "III"},
	{2, "II"},
	{1, "I"},
}
